A few small g++ related fixes.

+++ code/Wi11.cpp
@@ -240,7 +240,7 @@ RESULT Wi11::_Trap(const iWord& code) {
     } break;
     case 0x43: {
       Word new_R0_value;
-      new_R0_value.FromInt(rand()*pow(-1.0f, (float)rand())); // 2^16 = 65,536
+      new_R0_value.FromInt((int) (rand()*pow(-1.0f, (float)rand()))); // 2^16 = 65,536
       _R0 = new_R0_value;
 
       _UpdateCCR(_R0.GetValue().ToInt2Complement());
